zero-shot-evaluation/VLMEvalKit/requirements.txt
decord; platform_machine != 'arm64'
eva-decord; platform_machine == 'arm64'
gradio
huggingface_hub
imageio
matplotlib
numpy>=1.23.4
omegaconf
openai==1.3.5
opencv-python>=4.4.0.46
openpyxl
pandas
peft
pillow
portalocker
protobuf
python-dotenv
requests
rich
sentencepiece
setuptools
sty
tabulate
tiktoken
timeout-decorator
# torch>=2.0.1
tqdm
transformers
typing_extensions==4.7.1
validators
xlsxwriter
